# Conflicts:
#	common/ASC.Core.Common/EF/Context/BaseDbContext.cs
#	common/ASC.Data.Backup.Core/Service/ProgressItems/BackupProgressItem.cs
#	common/ASC.Data.Backup.Core/Storage/ConsumerBackupStorage.cs
#	common/ASC.Data.Backup.Core/Storage/DocumentsBackupStorage.cs
#	common/ASC.Data.Backup.Core/Tasks/BackupPortalTask.cs
#	common/ASC.Data.Backup.Core/Tasks/PortalTaskBase.cs
#	common/Tools/ASC.Migration.Creator/Program.cs
#	common/Tools/ASC.Migration.Runner/Program.cs
#	common/Tools/ASC.MigrationPersonalToDocspace/MigrationCreator.cs
#	common/Tools/ASC.MigrationPersonalToDocspace/MigrationRunner.cs

+++ common/ASC.Data.Storage/BaseStorage.cs
@@ -149,7 +149,7 @@ public abstract class BaseStorage : IDataStore
 
     public abstract Task<Stream> GetReadStreamAsync(string domain, string path);
 
-    public abstract Task<Stream> GetReadStreamAsync(string domain, string path, int offset);
+    public abstract Task<Stream> GetReadStreamAsync(string domain, string path, long offset);
 
     public abstract Task<Uri> SaveAsync(string domain, string path, Stream stream);
     public abstract Task<Uri> SaveAsync(string domain, string path, Stream stream, ACL acl);
