+++ common/ASC.Data.Storage/IDataStore.cs
@@ -95,7 +95,7 @@ public interface IDataStore
     ///<param name="domain"></param>
     ///<param name="path"></param>
     ///<returns></returns>
-    Task<Stream> GetReadStreamAsync(string domain, string path, int offset);
+    Task<Stream> GetReadStreamAsync(string domain, string path, long offset);
 
     ///<summary>
     /// Saves the contents of the stream in the repository.
